Make extractor simple
