+++ mars/services/meta/store/dictionary.py
@@ -15,7 +15,8 @@
 from dataclasses import asdict
 from typing import Dict, List
 
-from ....utils import implements, extensible
+from .... import oscar as mo
+from ....utils import implements
 from ....typing import BandType
 from ..core import _CommonMeta, _ChunkMeta
 from .base import AbstractMetaStore, register_meta_store
@@ -44,7 +45,7 @@ class DictMetaStore(AbstractMetaStore):
         self._store[object_id] = meta
 
     @implements(AbstractMetaStore.set_meta)
-    @extensible
+    @mo.extensible
     async def set_meta(self,
                        object_id: str,
                        meta: _CommonMeta):
@@ -73,7 +74,7 @@ class DictMetaStore(AbstractMetaStore):
                 return
 
     @implements(AbstractMetaStore.get_meta)
-    @extensible
+    @mo.extensible
     async def get_meta(self,
                        object_id: str,
                        fields: List[str] = None,
@@ -91,7 +92,7 @@ class DictMetaStore(AbstractMetaStore):
         del self._store[object_id]
 
     @implements(AbstractMetaStore.del_meta)
-    @extensible
+    @mo.extensible
     async def del_meta(self,
                        object_id: str):
         self._del_meta(object_id)
