+++ mars/tests/test_utils.py
@@ -362,108 +362,6 @@ def test_quiet_stdio():
 
 
 @pytest.mark.asyncio
-@pytest.mark.parametrize('use_async', [False, True])
-async def test_batch_decorator(use_async):
-    def _wrap_async(func):
-        async def _wrapped(*args, **kwargs):
-            return func(*args, **kwargs)
-        _wrapped.__name__ = func.__name__
-        return _wrapped if use_async else func
-
-    class TestClass:
-        def __init__(self):
-            self.arg_list = []
-            self.kwarg_list = []
-
-        @utils.extensible
-        @_wrap_async
-        def method1(self, *args, **kwargs):
-            raise NotImplementedError
-
-        @method1.batch
-        @_wrap_async
-        def method1(self, args_list, kwargs_list):
-            self.arg_list.extend(args_list)
-            self.kwarg_list.extend(kwargs_list)
-            return [len(self.kwarg_list)] * len(args_list)
-
-        @utils.extensible
-        @_wrap_async
-        def method2(self, *args, **kwargs):
-            self.arg_list.append(tuple(a - 1 for a in args))
-            self.kwarg_list.append({k: v - 1 for k, v in kwargs.items()})
-            return len(self.kwarg_list)
-
-        @utils.extensible
-        @_wrap_async
-        def method3(self, *args, **kwargs):
-            self.arg_list.append(tuple(a * 2 for a in args))
-            self.kwarg_list.append({k: v * 2 for k, v in kwargs.items()})
-            return len(self.kwarg_list)
-
-        @method3.batch
-        @_wrap_async
-        def method3(self, args_list, kwargs_list):
-            self.arg_list.extend([tuple(a * 2 + 1 for a in args) for args in args_list])
-            self.kwarg_list.extend([{k: v * 2 + 1 for k, v in kwargs.items()}
-                                    for kwargs in kwargs_list])
-            return [len(self.kwarg_list)] * len(args_list)
-
-    if use_async:
-        assert asyncio.iscoroutinefunction(TestClass.method1)
-
-    test_inst = TestClass()
-
-    ret = test_inst.method1.batch(
-        test_inst.method1.delay(12),
-        test_inst.method1.delay(10))
-    ret = await ret if use_async else ret
-    assert ret == [2, 2]
-    assert test_inst.arg_list == [(12,), (10,)]
-    assert test_inst.kwarg_list == [{}, {}]
-
-    test_inst.arg_list.clear()
-    test_inst.kwarg_list.clear()
-
-    ret = test_inst.method1(12, kwarg=34)
-    ret = await ret if use_async else ret
-    assert ret == 1
-    assert test_inst.arg_list == [(12,)]
-    assert test_inst.kwarg_list == [{'kwarg': 34}]
-
-    if sys.version_info[:2] > (3, 6):
-        test_inst = TestClass()
-        ret = test_inst.method2.batch(
-            test_inst.method2.delay(12, kwarg=34),
-            test_inst.method2.delay(10, kwarg=33)
-        )
-        ret = await ret if use_async else ret
-        assert ret == [1, 2]
-        assert test_inst.arg_list == [(11,), (9,)]
-        assert test_inst.kwarg_list == [{'kwarg': 33}, {'kwarg': 32}]
-
-        test_inst = TestClass()
-        ret = test_inst.method3(15, kwarg=56)
-        ret = await ret if use_async else ret
-        assert ret == 1
-        ret = test_inst.method3.batch(
-            test_inst.method3.delay(16, kwarg=57),
-            test_inst.method3.delay(17, kwarg=58)
-        )
-        ret = await ret if use_async else ret
-        assert ret == [3, 3]
-        assert test_inst.arg_list == [(30,), (33,), (35,)]
-        assert test_inst.kwarg_list == [{'kwarg': 112}, {'kwarg': 115}, {'kwarg': 117}]
-
-        if use_async:
-            test_inst = TestClass()
-            ret = await test_inst.method2.batch(
-                test_inst.method2.delay(12, kwarg=34),
-                test_inst.method2.delay(10, kawarg=33))
-            assert ret == [1, 2]
-
-
-@pytest.mark.asyncio
 @pytest.mark.skipif(sys.version_info[:2] < (3, 7),
                     reason='asyncio task timeout detector is not supported on python versions below 3.7')
 async def test_asyncio_task_timeout_detector():
